feat(package.json): update dependencies and devDependencies
The versions of typescript, vite, and vue-tsc are updated to the
latest versions. Also, a new dependency @vue/language-core is added.
This ensures that the project is using the latest features and
improvements from these libraries.

feat(tsconfig.app.json): add skipLibCheck to tsconfig.app.json

The skipLibCheck compiler option is added to the tsconfig.app.json
file. This option skips type checking of declaration files (.d.ts) to
reduce compilation time.